Experimenting with black hole round trips
Maybe: if deflection angle is greater Pi/2 < \phi <  Pi, this implies deflection from the direction of the player camera plane. For Pi < \phi < 3 * Pi / 2, this implies a decaying (exiting) orbit one revolution around the black hole, and so on. (At least, this should be true if the impact parameter is greater than the Schwarzschild radius, which it can be.)