Hypothetical "gravity mirror"
We normally consider the gravity lens equation within a deflection angle of -pi / 2 to pi / 2. Outside of this range, there is solid angle in the point-of-view to fill, and the impact parameter outside of this range is not required to be less than the Schwarzschild radius, (indicating decaying orbit of light, in-falling). This suggest decaying orbits out-falling, successive tangent ranges of width pi incrementally adding half-orbits around the black hole. Full integer orbits are simple enough, but n + 1/2 orbits would act like a "gravity mirror," bending light from behind the player halfway around the black hole and back to the player's eyes. (Incidentally, because of the variability in distance of orbit, I think time from emission to destination still appears the same across the various periods.)