Speculative Schwarzschild interior
Forgive me. I have simulation tools, and I want to run a "thought experiment": what if Schwarzschild coordinates are valid in the interior of the event horizon radius (this region presumed disconnected from external, but not necessarily) and the positive/negative flip of radius and time are meaningful in the "obvious" sense?

The singularity in space would instead appear as a singularity in time. I further assume that the direction of player time within the disconnected region is set by the entropy gradient. Therefore, our arrow of time is effectively reversed. (This gets us thinking about different closed and open "deaths" for the interior region, though, in knowledge of both a potentially maximum global entropy, set by the maximum size of the black hole, but alternatively also the interval from total evaporation to first formation.)

I've further assumed that the time reversal of Lemaître coordinates, with radius and time exchanged, are similarly valid as a description of the interior, while obviously driving toward usual Lemaître coordinates being disconnected from the interior. Personally, I think this is the shakiest part, but let's see what ultimately sticks.